Move military templates into their own class
