+++ Vic2ToHoI4/Source/HOI4World/HoI4Country.cpp
@@ -29,6 +29,7 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.*/
 #include "HoI4Focus.h"
 #include "HoI4Leader.h"
 #include "HoI4Localisation.h"
+#include "MilitaryMappings.h"
 #include "Names.h"
 #include "../Mappers/CountryMapping.h"
 #include "../Mappers/GovernmentMapper.h"
@@ -552,6 +553,7 @@ bool sufficientUnits(map<string, double>& units, map<string, string> subs,
         }
         return true;
 }
+
 
 void HoI4Country::convertArmies(const map<string, HoI4::UnitMap>& unitMap, const vector<HoI4::DivisionTemplateType>& divisionTemplates)
 {
